Added LED support to Serpent Mother.
This adds 1 pattern that works, "HeartBeat", and one that's still in the works,
"FromImage".

I can see 1 update in the UI, but then the UI freezes. I get the same for the
Poof Chase so I think this is not a regression but something about the UI.

This does not add protocol support, this is yet a TODO.

I also don't know what the protocol mapping should be. Please advice.